cmp: replace open-coded channel allocation by raw1394_channel_modify
The benefit is that it works on kernel 2.6.30 firewire-core and
libraw1394 2.0.3 without write permission to the IRM's device file.

If I read the code correctly, iec61883_cmp_normalize_output doesn't care
whether the channel was already allocated by somebody else; it only
cares that the channel is allocated.  That's what the new code does too.

Signed-off-by: Stefan Richter <stefanr@s5r6.in-berlin.de>